Move ComplianceOperator pipelines to compliance operator packages
