+++ central/complianceoperator/pipelines/complianceoperator/complianceoperatorinfo/pipeline.go
