+++ central/complianceoperator/pipelines/complianceoperator/complianceoperatorinfo/pipeline_test.go
