+++ central/complianceoperator/pipelines/complianceoperatorprofiles/pipeline.go
