+++ central/complianceoperator/pipelines/complianceoperatorprofiles/pipeline_test.go
