+++ central/complianceoperator/pipelines/complianceoperatorresults/pipeline.go
