+++ central/complianceoperator/pipelines/complianceoperatorrules/pipeline.go
