+++ central/complianceoperator/pipelines/complianceoperatorscans/pipeline.go
