+++ central/complianceoperator/pipelines/complianceoperatorscans/pipeline_test.go
