+++ central/complianceoperator/pipelines/complianceoperatorscansettingbinding/pipeline.go
