+++ central/complianceoperator/v2/pipelines/complianceoperatorprofilesv2/pipeline.go
