+++ central/complianceoperator/v2/pipelines/complianceoperatorprofilesv2/pipeline_test.go
