+++ central/complianceoperator/v2/pipelines/complianceoperatorrulesv2/pipeline.go
