+++ central/complianceoperator/v2/pipelines/complianceoperatorscansettingbindingsv2/pipeline.go
