+++ central/complianceoperator/v2/pipelines/complianceoperatorscansettingbindingsv2/pipeline_test.go
