+++ central/complianceoperator/v2/pipelines/complianceoperatorscansv2/pipeline.go
