+++ central/complianceoperator/v2/pipelines/complianceoperatorscansv2/pipeline_test.go
