+++ central/complianceoperator/v2/pipelines/complianceoperatorsuitesv2/pipeline_test.go
