+++ central/sensor/service/pipeline/all/factory.go
@@ -3,6 +3,18 @@ package all
 import (
 	"context"
 
+	"github.com/stackrox/rox/central/complianceoperator/pipelines/complianceoperator/complianceoperatorinfo"
+	"github.com/stackrox/rox/central/complianceoperator/pipelines/complianceoperatorprofiles"
+	"github.com/stackrox/rox/central/complianceoperator/pipelines/complianceoperatorresults"
+	"github.com/stackrox/rox/central/complianceoperator/pipelines/complianceoperatorrules"
+	"github.com/stackrox/rox/central/complianceoperator/pipelines/complianceoperatorscans"
+	"github.com/stackrox/rox/central/complianceoperator/pipelines/complianceoperatorscansettingbinding"
+	"github.com/stackrox/rox/central/complianceoperator/v2/pipelines/complianceoperatorprofilesv2"
+	"github.com/stackrox/rox/central/complianceoperator/v2/pipelines/complianceoperatorresultsv2"
+	"github.com/stackrox/rox/central/complianceoperator/v2/pipelines/complianceoperatorrulesv2"
+	"github.com/stackrox/rox/central/complianceoperator/v2/pipelines/complianceoperatorscansettingbindingsv2"
+	"github.com/stackrox/rox/central/complianceoperator/v2/pipelines/complianceoperatorscansv2"
+	"github.com/stackrox/rox/central/complianceoperator/v2/pipelines/complianceoperatorsuitesv2"
 	hashManager "github.com/stackrox/rox/central/hash/manager"
 	"github.com/stackrox/rox/central/sensor/service/pipeline"
 	"github.com/stackrox/rox/central/sensor/service/pipeline/alerts"
@@ -10,18 +22,6 @@ import (
 	"github.com/stackrox/rox/central/sensor/service/pipeline/clusterhealthupdate"
 	"github.com/stackrox/rox/central/sensor/service/pipeline/clustermetrics"
 	"github.com/stackrox/rox/central/sensor/service/pipeline/clusterstatusupdate"
-	"github.com/stackrox/rox/central/sensor/service/pipeline/complianceoperator/complianceoperatorinfo"
-	"github.com/stackrox/rox/central/sensor/service/pipeline/complianceoperatorprofiles"
-	"github.com/stackrox/rox/central/sensor/service/pipeline/complianceoperatorprofilesv2"
-	"github.com/stackrox/rox/central/sensor/service/pipeline/complianceoperatorresults"
-	"github.com/stackrox/rox/central/sensor/service/pipeline/complianceoperatorresultsv2"
-	"github.com/stackrox/rox/central/sensor/service/pipeline/complianceoperatorrules"
-	"github.com/stackrox/rox/central/sensor/service/pipeline/complianceoperatorrulesv2"
-	"github.com/stackrox/rox/central/sensor/service/pipeline/complianceoperatorscans"
-	"github.com/stackrox/rox/central/sensor/service/pipeline/complianceoperatorscansettingbinding"
-	"github.com/stackrox/rox/central/sensor/service/pipeline/complianceoperatorscansettingbindingsv2"
-	"github.com/stackrox/rox/central/sensor/service/pipeline/complianceoperatorscansv2"
-	"github.com/stackrox/rox/central/sensor/service/pipeline/complianceoperatorsuitesv2"
 	"github.com/stackrox/rox/central/sensor/service/pipeline/deploymentevents"
 	"github.com/stackrox/rox/central/sensor/service/pipeline/enhancements"
 	"github.com/stackrox/rox/central/sensor/service/pipeline/imageintegrations"
